Refactor position + imports
