+++ .pre-commit-config.yaml
@@ -1,0 +1,10 @@
+# repos:
+#   - repo: https://github.com/psf/black
+#     hooks:
+#       - id: black
+#         language_version: python3.8  # Should match the version of Python you're using
+
+#   - repo: https://github.com/pycqa/isort
+#     hooks:
+#       - id: isort
+#         language_version: python3.8  # Should match the version of Python you're using
