+++ app/cutout_handler/__init__.py
@@ -1,0 +1,3 @@
+from app.common import cutout_handler_stub, s3_handler_stub
+
+from .server import cutout_app
