+++ app/cutout_handler/segment.py
@@ -1,14 +1,19 @@
-import numpy as np
-import torch
-from segment_anything import sam_model_registry, SamPredictor
+from app.common import cutout_handler_stub
+
+cutout_handler_stub.cls()
 
 
 class Segmenter:
+    import numpy as np
+
     def __init__(
         self,
         sam_encoder_version: str,
         sam_checkpoint_path: str,
     ):
+        import torch
+        from segment_anything import SamPredictor, sam_model_registry
+
         self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
         self.sam = sam_model_registry[sam_encoder_version](
             checkpoint=sam_checkpoint_path
@@ -16,6 +21,8 @@ class Segmenter:
         self.sam_predictor = SamPredictor(self.sam)
 
     def segment(self, image: np.ndarray, xyxy: np.ndarray) -> np.ndarray:
+        import numpy as np
+
         self.sam_predictor.set_image(image)
         result_masks = []
         for box in xyxy:
